requirements-dev.txt
BeautifulSoup4
aiohttp
FastAPI
aiofiles
uvicorn
pytest-asyncio
pytest-xdist